video-generation-service/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
google-generativeai==0.8.3
gtts==2.5.3
moviepy==1.0.3
pillow==10.4.0
imageio==2.35.1
imageio-ffmpeg==0.5.1
numpy==2.1.1
python-dotenv==1.0.1
requests==2.32.3
python-multipart==0.0.20
pydantic==2.12.3